+++ Src/MoneyFox.Ui/Views/Payments/PaymentModification/EditPaymentViewModel.cs
@@ -7,8 +7,8 @@ using Controls.CategorySelection;
 using Core.Common.Extensions;
 using Core.Common.Interfaces;
 using Core.Common.Settings;
-using Core.Features._Legacy_.Payments.DeletePaymentById;
 using Core.Features._Legacy_.Payments.UpdatePayment;
+using Core.Features.PaymentByIdDeletions;
 using Core.Queries.PaymentDataById;
 using MediatR;
 using Resources.Strings;
